requirements.txt.txt
streamlit
google-generativeai
python-dotenv
numpy